ethernet/nsfw/requirements.txt
appwrite
tensorflow
opennsfw2
pillow